test(mobile): add 47 new tests + fix iOS deserialization (#234)
* test(mobile): add 47 new tests across shared + android modules

- Add IProcessApi interface for testability, ProcessApi implements it
- ProcessingViewModel depends on IProcessApi instead of concrete class
- Add MetricsApiTest (5 tests): registry, trend, PRs, diagnostics, summary
- Add UsersApiTest (3 tests): getMe, updateProfile, updateSettings
- Add ProcessingViewModelTest (5 tests): progress→completed, queue failure,
  stream failed, cancel success/failure
- Add SerializationTest: +5 model tests (SessionUpdateRequest,
  MetricsRegistryResponse, TrendResponse, PRsResponse, DiagnosticsResponse)
- Add ExportSessionUseCaseTest (3 tests): export zip, manifest, error
- Add StreamingUseCasesTest (5 tests): start/stop streaming, factory reset
- Fix ProcessApi @OptIn(ExperimentalSerializationApi) warning

* fix(mobile): change UserResponse heightCm/weightKg from Float to Double

Kotlin/Native (iOS) throws IllegalArgumentException when deserializing
JSON numbers (always Double) into Float fields. Also add
@OptIn(ExperimentalCoroutinesApi::class) to ViewModel tests.

* test(mobile): add ProfileViewModel + AccCalibrate tests, fix Float→Double in SerializationTest

* test(mobile): add ProfileViewModel and ChunkedUploader tests

* test(mobile): add ImuStreamWriter edge case tests

* test(mobile): add ProcessApi queue with params test

* fix(test): remove MockEngine from ChunkedUploaderTest (not in androidApp test deps)

* fix(mobile): replace mapOf+setBody with @Serializable classes and buildJsonObject

Kotlin/Native iOS crashes with IllegalArgumentException: Star projections
in type arguments are not allowed when using mapOf() + setBody(). The
dynamic Map<String, Any?> type fails Kotlin/Native strict generics checks.

- AuthApi: mapOf → @Serializable data classes (LogoutRequest, VerifyEmailRequest,
  ResendVerificationRequest, ForgotPasswordRequest, ResetPasswordRequest)
- UsersApi: mapOf → buildJsonObject with explicit types
- SessionsApi: mapOf → buildJsonObject (create), @Serializable BulkDeleteRequest
- SkateLabClient: mapOf → buildJsonObject (refresh token)
- AuthApiTest: add serialization tests for new request classes

* test(mobile): add ViewModel tests and fix ChunkedUploaderTest OOM

+++ mobile/androidApp/src/main/java/ru/skatelab/capture/ui/profile/ProfileViewModel.kt
@@ -65,7 +65,7 @@ class ProfileViewModel
             displayName: String? = null,
             bio: String? = null,
             heightCm: Int? = null,
-            weightKg: Float? = null,
+            weightKg: Double? = null,
         ) {
             viewModelScope.launch {
                 _uiState.value = _uiState.value.copy(isSaving = true, saveSuccess = false)
